agent/anomaly/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "anomaly-agent"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "bedrock-agentcore >= 1.0.3",
    "httpx >= 0.27.0",
    "pyyaml >= 6.0",
    "strands-agents >= 1.13.0",
]
